Moved all types into types
