+++ src/index.ts
@@ -13,14 +13,17 @@ import {
 export type {
     ComponentDefinition,
     ContentFetcher,
+    ContentPathItem,
     Context,
     FetchContentResult,
     GuillotineResult,
+    MacroProps,
     MetaData,
     PageComponent,
     PageData,
     PageRegion,
     PartData,
+    PartProps,
     QueryGetter,
     RegionTree,
     SelectedQueryMaybeVariablesFunc,
@@ -30,30 +33,45 @@ export type {
 
 
 export {ComponentRegistry} from './ComponentRegistry';
+
 export {
+    JSESSIONID_HEADER,
     PORTAL_COMPONENT_ATTRIBUTE,
+    PROJECT_ID_HEADER,
     RENDER_MODE,
+    RENDER_MODE_HEADER,
     XP_BASE_URL_HEADER,
+    XP_COMPONENT_TYPE,
+    XP_REQUEST_TYPE,
 } from './constants';
+
 export {
     APP_NAME,
     IS_DEV_MODE,
 } from './env';
+
 export {fetchContent} from './guillotine/fetchContent';
 export {fetchContentPathsForAllLocales} from './guillotine/fetchContentPathsForAllLocales';
+export {fetchContentPathsForLocale} from './guillotine/fetchContentPathsForLocale';
 export {fetchFromApi} from './guillotine/fetchFromApi';
 export {fetchGuillotine} from './guillotine/fetchGuillotine';
 export {richTextQuery} from './guillotine/metadata/richTextQuery';
 export {validateData} from './guillotine/validateData';
+
+export {I18n} from './i18n/i18n';
+export {LocaleContextProvider} from './i18n/LocaleContext';
+
+export {getContentApiUrl} from './utils/getContentApiUrl';
+export {getProjectLocaleConfig} from './utils/getProjectLocaleConfig';
+export {getProjectLocaleConfigById} from './utils/getProjectLocaleConfigById';
 export {getRequestLocaleInfo} from './utils/getRequestLocaleInfo';
 export {sanitizeGraphqlName} from './utils/sanitizeGraphqlName';
+
 export {
     UrlProcessor,
     getAsset,
     getUrl,
 } from './UrlProcessor';
-export {I18n} from './i18n/i18n';
-export {LocaleContextProvider} from './i18n/LocaleContext';
 
 
 const adapterConstants = {
