+++ src/views/macros/DisableMacro.tsx
@@ -1,5 +1,8 @@
+import type {MacroProps} from '../../types';
+
+
 import React from 'react';
-import {MacroProps} from '../BaseMacro';
+
 
 const DisableMacro = ({name, config, meta}: MacroProps) => (
     config.body
